refactor(session-notification): make sender a pure facade
